Introduce SchemaKey.PARAMETER_ORDER_PRIORITY and move order_schema_properties to utils/pydantic
