+++ obi_one/scientific/tasks/generate_simulations/config/brian2/brian2_circuit.py
@@ -72,7 +72,7 @@ class Brian2CircuitSimulationScanConfig(Brian2SimulationScanConfig):
             description="Circuit to simulate.",
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.MODEL_IDENTIFIER,
-                "order_priority": 100,
+                SchemaKey.PARAMETER_ORDER_PRIORITY: 100,
             },
         )
         node_set: NeuronSetReference | None = Field(
@@ -82,7 +82,7 @@ class Brian2CircuitSimulationScanConfig(Brian2SimulationScanConfig):
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.REFERENCE,
                 SchemaKey.REFERENCE_TYPE: NeuronSetReference.__name__,
-                "order_priority": 99,
+                SchemaKey.PARAMETER_ORDER_PRIORITY: 99,
             },
         )
