+++ obi_one/scientific/tasks/generate_simulations/config/neuron/neuron_me_model.py
@@ -44,7 +44,7 @@ class MEModelSimulationScanConfig(NeuronSimulationScanConfig):
             description="ME Model to simulate.",
             json_schema_extra={
                 SchemaKey.UI_ELEMENT: UIElement.MODEL_IDENTIFIER,
-                "order_priority": 100,
+                SchemaKey.PARAMETER_ORDER_PRIORITY: 100,
             },
         )
